IGNITE-24850 Split IgniteUtils to common and core specific methods
