+++ modules/commons/src/main/java/org/apache/ignite/internal/util/CacheClassLoaderMarker.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.ignite.internal.processors.cache;
+package org.apache.ignite.internal.util;
 
 /**
  * It's a marker interface for detecting GridCacheDeploymentManager$CacheClassLoader.
