+++ modules/commons/src/main/java/org/apache/ignite/internal/util/ClassCache.java
